Fixes and updates

+++ Argus.Automation.Example/Restaurant.Tests/Helpers/RoundHelper.cs
@@ -2,7 +2,7 @@
 {
     public static class RoundHelper
     {
-        public static double DefaultRound(this double value, int roundDigits = 2)
+        public static decimal DefaultRound(this decimal value, int roundDigits = 2)
         {
             return Math.Round(value, roundDigits, MidpointRounding.AwayFromZero);
         }
